fix update

+++ FinalProject/src/main/java/com/hit/controller/ActivityController.java
@@ -114,7 +114,7 @@ public class ActivityController {
 
 				LOGGER.info("user is logged in");
 
-				String activityName = request.getParameter("activityName");
+				String activityName = request.getParameter("exercise_name");
 				String activityDate = request.getParameter("activityDate");
 
 				String userName = (String) session.getAttribute("userName");
@@ -122,47 +122,54 @@ public class ActivityController {
 				if (inputValidator.activityUpdateValidation(request)) {
 					activity = hibernateGymDAO.getActivity(userName, activityName, activityDate);
 
-					if (activityName.equals("cardio")) {
-						if (request.getParameter("amount_of_sets") != "" || request.getParameter("repeats") != "" || request.getParameter("weight") != "") {
-							request.setAttribute("message", "You cannot change amount of sets, repeats or weight in cardio activity");
-							request.getRequestDispatcher("/updateActivity.jsp").forward(request, response);
-							return;
+					if (activity != null) {
+
+						if (activityName.equals("cardio")) {
+							if (request.getParameter("amount_of_sets") != "" || request.getParameter("repeats") != ""
+									|| request.getParameter("weight") != "") {
+								request.setAttribute("message",
+										"You cannot change amount of sets, repeats or weight in cardio activity");
+								request.getRequestDispatcher("/updateActivity.jsp").forward(request, response);
+								return;
+							}
+							if (request.getParameter("duration") != "") {
+								activity.setDuration(Float.parseFloat(request.getParameter("duration")));
+							}
+						} else {
+							if (request.getParameter("duration") != "") {
+								request.setAttribute("message", "You cannot change duration in anaerobic activity");
+								request.getRequestDispatcher("/updateActivity.jsp").forward(request, response);
+								return;
+							}
+							if (request.getParameter("amount_of_sets") != "") {
+								activity.setAmountOfSets(Integer.parseInt(request.getParameter("amount_of_sets")));
+							}
+
+							if (request.getParameter("repeats") != "") {
+								activity.setAmountOfRepeatition(Integer.parseInt(request.getParameter("repeats")));
+							}
+
+							if (request.getParameter("weight") != "") {
+								activity.setWeight(Float.parseFloat(request.getParameter("weight")));
+							}
 						}
-						if (request.getParameter("duration") != "") {
-							activity.setDuration(Float.parseFloat(request.getParameter("duration")));
+
+						// Validation succeeded, updating attempt:
+						LOGGER.info("Input is valid, trying to update.");
+						if (hibernateGymDAO.updateActivity(activity)) {
+							LOGGER.info("Activity has been updated");
+							request.setAttribute("message", "Activity has been updated successfully");
+							request.getRequestDispatcher("/activities.jsp").forward(request, response);
+
+						} else {
+							LOGGER.info("Activity cannot be updated");
+							request.setAttribute("message", "Activity cannot be updated");
+							request.getRequestDispatcher("/updateActivity.jsp").forward(request, response);
 						}
 					} else {
-						if (request.getParameter("duration") != "") {
-							request.setAttribute("message", "You cannot change duration in anaerobic activity");
-							request.getRequestDispatcher("/updateActivity.jsp").forward(request, response);
-							return;
-						}
-						if (request.getParameter("amount_of_sets") != "") {
-							activity.setAmountOfSets(Integer.parseInt(request.getParameter("amount_of_sets")));
-						}
-
-						if (request.getParameter("repeats") != "") {
-							activity.setAmountOfRepeatition(Integer.parseInt(request.getParameter("repeats")));
-						}
-
-						if (request.getParameter("weight") != "") {
-							activity.setWeight(Float.parseFloat(request.getParameter("weight")));
-						}
-					}
-
-					// Validation succeeded, updating attempt:
-					LOGGER.info("Input is valid, trying to update.");
-					if (hibernateGymDAO.updateActivity(activity)) {
-						LOGGER.info("Activity has been updated");
-						request.setAttribute("message", "Activity has been updated successfully");
-						request.getRequestDispatcher("/activities.jsp").forward(request, response);
-
-					} else {
-						LOGGER.info("Activity cannot be updated");
-						request.setAttribute("message", "Activity cannot be updated");
+						request.setAttribute("message", "Activity dose not exit");
 						request.getRequestDispatcher("/updateActivity.jsp").forward(request, response);
 					}
-
 				} else {
 					// Validation fails:
 					LOGGER.info("Activity is not valid");
@@ -199,7 +206,7 @@ public class ActivityController {
 
 				LOGGER.info("user is logged in");
 
-				String activityName = request.getParameter("activityName");
+				String activityName = request.getParameter("exercise_name");
 				String activityDate = request.getParameter("activityDate");
 
 				String userName = (String) session.getAttribute("userName");
@@ -208,17 +215,26 @@ public class ActivityController {
 
 					// Validation succeeded, deleting attempt:
 					LOGGER.info("Input is valid, going to update");
-					if (hibernateGymDAO.deleteActivity(userName, activityName, activityDate)) {
-						LOGGER.info("Activity has been deleted");
-						request.setAttribute("message", "Activity deleted successfully");
-						request.getRequestDispatcher("/activities.jsp").forward(request, response);
+
+					Activity activity = hibernateGymDAO.getActivity(userName, activityName, activityDate);
+
+					if (activity != null) {
+
+						if (hibernateGymDAO.deleteActivity(userName, activityName, activityDate)) {
+							LOGGER.info("Activity has been deleted");
+							request.setAttribute("message", "Activity deleted successfully");
+							request.getRequestDispatcher("/activities.jsp").forward(request, response);
+
+						} else {
+							LOGGER.info("Activity cannot be deleted");
+							request.setAttribute("message", "Activity cannot be deleted");
+							request.getRequestDispatcher("/deleteActivity.jsp").forward(request, response);
+						}
 
 					} else {
-						LOGGER.info("Activity cannot be deleted");
-						request.setAttribute("message", "Activity cannot be deleted");
+						request.setAttribute("message", "Activity dose not exit");
 						request.getRequestDispatcher("/deleteActivity.jsp").forward(request, response);
 					}
-
 				} else {
 					// Validation fails:
 					LOGGER.info("Activity is not valid");
